Improve is_new_doc calculation.
In `BP_Docs_Query::save()`, make the `is_new_logic` be less context sensitive. The new logic assumes the doc is new if the save args didn't specify a doc_id to update or the doc_id was passed, and it refers to post with the status set to `auto-draft`.