Allow folders to be enabled in a single context only.
Change the admin setting and supporting utility functions to accept and return a single context only.
Repurpose `bp_docs_get_folder_setting()` to be an option fetcher and sanitizer only and introduce `bp_docs_get_folders_enabled_for_context()` to return whether or not folders should be enabled for a particular context.