Change logic in bp_docs_folder_tax_query().
This function affects what is shown "outside" of any folders in a view. Previously, what was shown was only items that were not in any folder. This change shows items that are not in folders that are associated with the current context. For instance, if you are viewing the "Started by me" tab, you should see docs that you started that are in a global folder or a group folder.